Corrections solver humain exploration
